+++ ydb/tests/olap/column_compression/create/set_compression.py
@@ -1,9 +1,9 @@
 import logging
-from .base import ColumnTestBase
-from ydb.tests.library.common.helpers import plain_or_under_sanitizer
-from ydb.tests.olap.common.column_table_helper import ColumnTableHelper
 import pytest
 import random
+from ydb.tests.olap.column_compression.common.base import ColumnTestBase
+from ydb.tests.library.common.helpers import plain_or_under_sanitizer
+from ydb.tests.olap.common.column_table_helper import ColumnTableHelper
 
 logger = logging.getLogger(__name__)
