+++ ydb/tests/olap/column_compression/create/ya.make
@@ -4,7 +4,6 @@ INCLUDE(${ARCADIA_ROOT}/ydb/tests/harness_dep.inc)
 FORK_SUBTESTS()
 
 TEST_SRCS(
-    base.py
     set_compression.py
 )
 
@@ -16,6 +15,7 @@ PEERDIR(
     ydb/public/sdk/python/enable_v3_new_behavior
     ydb/tests/olap/scenario/helpers
     ydb/tests/olap/common
+    ydb/tests/olap/column_compression/common
 )
 
 DEPENDS(
