+++ ydb/tests/olap/column_compression/fast/test_obsolete.py
